requirements.txt
streamlit
streamlit-webrtc
requests
google-cloud-storage
pandas
numpy
matplotlib
seaborn
streamlit_lottie
htbuilder
